requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.4
redis>=5.0.0
celery>=5.3.0
django-environ>=0.11.0
django-filter>=24.1
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0
Pillow>=10.2.0